# Unroll skipping loops to mitigate slowdown wiht GraalVM