Below are results of a benchmark that is shared by @PavelPenkov [here](https://github.com/PavelPenkov/debezium_json), for more details see [his message on reddit](https://www.reddit.com/r/scala/comments/cjx9r0/the_best_json_parser_with_dynamic_types/eww4io2/)

## Before
```
[info] Running (fork) org.openjdk.jmh.Main -i 2 -wi 2 -f 1
[info] # JMH version: 1.21
[info] # VM version: JDK 1.8.0_222, OpenJDK 64-Bit GraalVM CE 19.1.1, 25.222-b08-jvmci-19.1-b01
[info] # VM invoker: /usr/lib/jvm/graalvm-ce-19/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 2 iterations, 10 s each
[info] # Measurement: 2 iterations, 10 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: ru.example.DebeziumJson.jsonIter
[info] # Run progress: 0.00% complete, ETA 00:00:40
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: 24437.675 ops/s
[info] # Warmup Iteration   2: 24435.290 ops/s
[info] Iteration   1: 24466.848 ops/s
[info] Iteration   2: 24073.207 ops/s
[info] Result "ru.example.DebeziumJson.jsonIter":
[info]   24270.028 ops/s
[info] # Run complete. Total time: 00:00:40
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark               Mode  Cnt      Score   Error  Units
[info] DebeziumJson.jsonIter  thrpt    2  24270.028          ops/s
```

## After
```
[info] Running (fork) org.openjdk.jmh.Main -i 2 -wi 2 -f 1
[info] # JMH version: 1.21
[info] # VM version: JDK 1.8.0_222, OpenJDK 64-Bit GraalVM CE 19.1.1, 25.222-b08-jvmci-19.1-b01
[info] # VM invoker: /usr/lib/jvm/graalvm-ce-19/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 2 iterations, 10 s each
[info] # Measurement: 2 iterations, 10 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: ru.example.DebeziumJson.jsonIter
[info] # Run progress: 0.00% complete, ETA 00:00:40
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: 28992.175 ops/s
[info] # Warmup Iteration   2: 28888.802 ops/s
[info] Iteration   1: 29174.056 ops/s
[info] Iteration   2: 29062.905 ops/s
[info] Result "ru.example.DebeziumJson.jsonIter":
[info]   29118.481 ops/s
[info] # Run complete. Total time: 00:00:40
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark               Mode  Cnt      Score   Error  Units
[info] DebeziumJson.jsonIter  thrpt    2  29118.481          ops/s
```